refactor: single file for OpenTelemetryXxxCollectorFilterAttribute
